requirements.txt
fastapi
uvicorn
requests
faiss-cpu
pydantic
python-multipart